\name{vampyr-internal}
\alias{vampyr-internal}
\alias{eap_continuous_obli}
\alias{eap_grad_obli}
\alias{Nseriesk}
\alias{ordnormulti}
\alias{pconmul}
\alias{ppnd}
\alias{reap_grad_obli}
\alias{thresholds}
\title{
vampyr-internal
}
\description{
This functions are for internal use only.
}
\usage{
eap_continuous_obli
eap_grad_obli
Nseriesk
ordnormulti
pconmul
ppnd
reap_grad_obli
thresholds
}
\author{
David Navarro-Gonzalez

Andreu Vigil-Colet

Pere Joan Ferrando

Urbano Lorenzo-Seva

Jorge Tendeiro
}
\keyword{ internal }
