\name{vampyr-package}
\alias{vampyr-package}
\title{
Factor Analysis Controlling the Effects of Response Bias
}
\description{
Package for performing Factor Analysis (FA) implementing the method developed by Ferrando, Lorenzo-Seva & Chico (2009). It allows the user to control the effects of two response bias: Social Desirability (SD) and Acquiescence (AC) by extracting the variance due to these factors before extracting the content ones. In other words, vampirize the response biases by substracting their impact on the factor structure.
}
\value{
\item{\code{\link{ControlResponseBias}}}{Performs FA controlling response biases.}
}
\references{
Ferrando, P. J., Lorenzo-Seva, U., & Chico, E. (2009). A General factor-Analytic procedure for assessing response bias in questionnaire measures. Structural Equation Modeling, 16, 364-381. \doi{10.1080/10705510902751374}
}
\author{
David Navarro-Gonzalez

Andreu Vigil-Colet

Pere Joan Ferrando

Urbano Lorenzo-Seva

Jorge Tendeiro
}
\keyword{ package }
