\name{vampyr_example}
\alias{vampyr_example}
\docType{data}
\title{
vampyr_example database
}
\description{
A database to be used as example in the functions included on \code{controlSDAQ} package. It contains the answers of 300 participants to the Physical aggression subscale of IDAQ questionnaire (Ruiz-Pamies, Lorenzo-Seva, Morales-Vives, Cosi, Vigil-Colet, 2014), which was developed for assessing Physical, Verbal and Indirect aggression. The originsl questionnaire contains 27 Likert-items, ranging from 1 to 5, and this dataset only contains the 4 SD markers and the 6 Physical aggression items.
}
\usage{data("vampyr_example")}
\format{
  A data frame with 300 observations and 10 variables, where 6 items measuring Physical aggression and 4 Social Desirability markers.
}
\details{
The original sample contains 750 participants, and the following database only contains 300 selected at random.
}
\source{
More information about the questionnaire can be found at:

\url{http://psico.fcep.urv.cat/tests/idaq/en/descripcion.html}
}
\references{
Ruiz-Pamies, M., Lorenzo-Seva, U., Morales-Vives, F., Cosi, S., & Vigil-Colet, A. (2014). I-DAQ: a new test to assess direct and indirect aggression free of response bias. The Spanish Journal of Psychology, 17, E41. \url{http://doi.org/10.1017/sjp.2014.43}
}
\examples{
data(vampyr_example)
}
\keyword{datasets}
